A4A WooPayments: extract dashboard data orchestration into a hook
Move the WooPayments commissions dashboard's data orchestration (WooPayments
licenses and plugin sites, connection health, and the commissions data) out of
the component into a useWooPaymentsDashboardData hook, so the component is just
layout and rendering. Behavior unchanged.
